feat(platform): make billing and routing contracts decimal
Use decimal for balances, quotas, rate multipliers, daily limits, rate windows, and usage cost across Orleans interfaces, state, Admin API DTOs, and endpoint calls. Preserve exact decimal values through auth, group, account, user, and API-key projections without intermediate binary floating-point conversion.

Keep explicit double casts only at the current Cap'n Proto Float64 serialization boundary and document that fixed-scale/text contract regeneration is a later package. Add a high-precision cross-projection regression test and update grain/benchmark fixtures to decimal literals.

Verification: dotnet build ScalaAPI.Platform.slnx --no-restore -c Release (0 warnings, 0 errors); dotnet test ScalaAPI.Platform.slnx --no-build -c Release (63 passed).

+++ src/Admin.Api/Endpoints/PlatformEndpoints.cs
@@ -229,7 +229,7 @@ public static class PlatformEndpoints
                         UserId = payment.UserId, PaymentId = payment.Id,
                         Reference = $"payment:{payment.Id}", Amount = payment.Amount,
                     }).ExecuteCommandAsync();
-                    await client.GetGrain<IUserGrain>(payment.UserId).AdjustBalance((double)payment.Amount);
+                    await client.GetGrain<IUserGrain>(payment.UserId).AdjustBalance(payment.Amount);
                 }
                 catch (Exception)
                 {
@@ -334,7 +334,7 @@ public static class PlatformEndpoints
                 {
                     UserId = user.Id, Reference = $"redeem:{code.Id}:{user.Id}", Amount = code.BonusAmount,
                 }).ExecuteCommandAsync();
-                await client.GetGrain<IUserGrain>(user.Id).AdjustBalance((double)code.BonusAmount);
+                await client.GetGrain<IUserGrain>(user.Id).AdjustBalance(code.BonusAmount);
             }
             return Results.Ok(new { message = "Code redeemed", bonus = code.BonusAmount });
         });
@@ -745,9 +745,9 @@ public static class PlatformEndpoints
     }
 
     private record ApiKeySelfServiceRequest(
-        string? Name, long? GroupId, double? Quota, long? ExpiresAt,
+        string? Name, long? GroupId, decimal? Quota, long? ExpiresAt,
         string[]? IpWhitelist, string[]? IpBlacklist,
-        double? RateLimit5h, double? RateLimit1d, double? RateLimit7d);
+        decimal? RateLimit5h, decimal? RateLimit1d, decimal? RateLimit7d);
     private record RedeemRequest(string Code);
     private record ChannelCheckRequest(long AccountId, string Status, int LatencyMs, string? Error);
     private record RestoreRequest(string BackupId);
